Create common module for mock BootHostManager
